% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGT3X.R
\name{as.vector.activity}
\alias{as.vector.activity}
\title{Activity matrix as vector magnitude vector}
\usage{
\method{as.vector}{activity}(activity)
}
\arguments{
\item{activity}{Matrix with class 'activity' as returned by read.gt3x}
}
\value{
A vector with class 'activity_vm' which has length equal to nrow(activity).
Each index holds the vector magnitude (sqrt(x^2 + y^2 + z^2)) of the corresponding input row.
The object has the following attributes:
\itemize{
\item \code{subject_name} : Subject name from info file
\item \code{time_zone} : Time zone from info file
\item \code{missingness} : Data frame with timestamps and the number of missing values associated.
\item \code{time} : timestamp associated with each sample
}
}
\description{
Activity matrix as vector magnitude vector
}
