% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_sample_data.R
\name{gt3x_dataurl}
\alias{gt3x_dataurl}
\title{Get url of github release}
\usage{
gt3x_dataurl(
  version = "v1.0",
  baseurl = "https://github.com/THLfi/read.gt3x/releases/download"
)
}
\arguments{
\item{version}{release version}

\item{baseurl}{URL for GitHub release}
}
\value{
URL to file
}
\description{
Get url of github release
}
\seealso{
Other sample-data:
\code{\link[=gt3x_download]{gt3x_download()}},
\code{\link[=gt3x_url]{gt3x_url()}}
}
\concept{sample-data}
