% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_sample_data.R
\name{gt3x_download}
\alias{gt3x_download}
\title{Download and unzip a zipped \code{gt3x} file}
\usage{
gt3x_download(url, exdir, verbose = TRUE)
}
\arguments{
\item{url}{url of the file to download}

\item{exdir}{directory to extract the zip file}

\item{verbose}{print diagnostic messages}
}
\value{
file path of \code{exdir}
}
\description{
Download and unzip a zipped \code{gt3x} file
}
\seealso{
Other sample-data:
\code{\link[=gt3x_dataurl]{gt3x_dataurl()}},
\code{\link[=gt3x_url]{gt3x_url()}}
}
\concept{sample-data}
