% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{have_log_and_info}
\alias{have_log_and_info}
\title{Check if a .gt3x file or unzipped gt3x directory has both log.bin adn info.txt}
\usage{
have_log_and_info(gt3x)
}
\description{
Check if a .gt3x file or unzipped gt3x directory has both log.bin adn info.txt
}
\seealso{
Other gt3x-utils: \code{\link{is_gt3x}}
}
\concept{gt3x-utils}
