% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parseGT3X}
\alias{parseGT3X}
\title{Parse activity samples from a GT3X file}
\usage{
parseGT3X(filename, max_samples, scale_factor, sample_rate,
  verbose = FALSE, debug = FALSE, impute_zeroes = FALSE)
}
\arguments{
\item{filename}{(char*) path to a log.bin file inside the unzipped gt3x folder, which contains the activity samples}

\item{max_samples}{Maximum number of rows to parse. The returned matrix will always contain this number of rows, having zeroes if
not data is found.}

\item{verbose}{Print the parameters from the log.bin file and other messages?}

\item{debug}{Print information for every activity second}

\item{scale}{factor Scale factor for the activity samples.}
}
\value{
Returns a matrix with max_samples rows and 4 columns, where the first 3 columns are the acceleration samples and
the last column is timestamps in seconds (including 100th of seconds) starting from 00:00:00 1970-01-01 UTC (UNIX time)
}
\description{
Parse activity samples from a GT3X file
}
