% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parseLuxBin}
\alias{parseLuxBin}
\title{Parse activity samples from a GT3X file}
\usage{
parseLuxBin(filename, max_samples, scale_factor, max_value, verbose = FALSE)
}
\arguments{
\item{filename}{(char*) path to a log.bin file inside the unzipped gt3x folder, which contains the activity samples}

\item{max_samples}{Maximum number of rows to parse. The returned matrix will always contain this number of rows, having zeroes if
not data is found.}

\item{scale_factor}{Scale factor for the activity samples.}

\item{max_value}{Maximum value to truncate}

\item{verbose}{Print the parameters from the log.bin file and other messages?}
}
\value{
Returns a vector with max_samples elements
}
\description{
Parse activity samples from a GT3X file
}
