% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGT3X.R
\name{print.activity_df}
\alias{print.activity_df}
\alias{head.activity_df}
\alias{print.activity}
\alias{head.activity}
\title{Print the contents of the activity data}
\usage{
\method{print}{activity_df}(x, ...)

\method{head}{activity_df}(x, ...)

\method{print}{activity}(x, ...)

\method{head}{activity}(x, ...)
}
\arguments{
\item{x}{gt3x_info object returned by parse_gt3x_info()}

\item{...}{additional arguments passed to \code{\link{head}}}
}
\description{
Print the contents of the activity data
}
