% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGT3X.R
\name{readGT3X}
\alias{readGT3X}
\title{Read GT3X}
\usage{
readGT3X(path, verbose = FALSE, ...)
}
\arguments{
\item{path}{Path to gt3x folder}
}
\description{
Read activity samples from a GT3X file as a matrix
}
\examples{

# first unzip, then read
datadir <- gt3x_datapath()
gt3xfolders <- unzip.gt3x(datadir)
x <- readGT3X(gt3xfolders[1])
df <- as.data.frame(x)

# temporary unzip and read
gt3xfile <- gt3x_datapath(1)
x <- readGT3X(gt3xfile)

}
\seealso{
Other gt3x-parser: \code{\link{as.data.frame.activity}},
  \code{\link{print.gt3x_info}}
}
\concept{gt3x-parser}
