% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGT3X.R
\name{as.data.frame.activity}
\alias{as.data.frame.activity}
\title{Convert an activity matrix to a data.frame}
\usage{
\method{as.data.frame}{activity}(activity)
}
\arguments{
\item{activity}{Object of class 'activity' (returned by readGT3X)}
}
\description{
Convert an activity matrix to a data.frame
}
\seealso{
Other gt3x-parser: \code{\link{print.gt3x_info}},
  \code{\link{readGT3X}}
}
\concept{gt3x-parser}
