% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_zeros.R
\name{fill_zeros}
\alias{fill_zeros}
\alias{fill_zeroes}
\title{Fill Zeroes to Last Observation Carried Forward}
\usage{
fill_zeros(x)

fill_zeroes(x)
}
\arguments{
\item{x}{Object of class \code{activity}/\code{activity_df} or a \code{data.frame},
with columns \code{X}/\code{Y}/\code{Z}}
}
\value{
A similar object passed in with XYZ filled in with last observation
}
\description{
Fill Zeroes to Last Observation Carried Forward
}
