% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGT3X.R
\name{get_n_samples}
\alias{get_n_samples}
\title{Calculate the expected activity sample size from start time and last sample time in the info.txt of a gt3x directory}
\usage{
get_n_samples(info)
}
\description{
Calculate the expected activity sample size from start time and last sample time in the info.txt of a gt3x directory
}
