% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_sample_data.R
\name{gt3x_datapath}
\alias{gt3x_datapath}
\alias{gt3x_download}
\alias{gt3x_url}
\title{Path to read.gt3x package sample data}
\usage{
gt3x_datapath(index = NULL, verbose = TRUE)

gt3x_download(url, exdir, verbose = TRUE)

gt3x_url(index = NULL, filename = NULL)
}
\arguments{
\item{index}{Integer. The index of a sample file to retrieve. If NULL (default) the path to the directory
including the sample files will be returned.}

\item{verbose}{print diagnostic messages}

\item{url}{url of the file to download}

\item{exdir}{directory to extract the zip file}

\item{filename}{file to grab to make url}

\item{verbose}{print diagnostic messages}
}
\description{
Path to read.gt3x package sample data

Download and unzip a zipped gt3xfile

Get url of gt3x sample file
}
\examples{
dir <- gt3x_datapath()
gt3x_filename <- gt3x_datapath(1)

}
\seealso{
Other file manipulations: \code{\link{is_gt3x}},
  \code{\link{list_gt3x}}, \code{\link{unzip.gt3x}}

Other sample-data: \code{\link{gt3x_dataurl}}

Other sample-data: \code{\link{gt3x_dataurl}}
}
\concept{file manipulations}
\concept{sample-data}
