% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_sample_data.R
\name{gt3x_datapath}
\alias{gt3x_datapath}
\title{Path to read.gt3x package sample data}
\usage{
gt3x_datapath(index = NULL)
}
\arguments{
\item{index}{Integer. The index of a sample file to retrieve. If NULL (default) the path to the directory
including the sample files will be returned.}
}
\description{
Path to read.gt3x package sample data
}
\examples{
dir <- gt3x_datapath()
gt3x_filename <- gt3x_datapath(1)

}
