% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_sample_data.R
\name{gt3x_download}
\alias{gt3x_download}
\title{Download and unzip a zipped gt3xfile}
\usage{
gt3x_download(url, exdir)
}
\description{
Download and unzip a zipped gt3xfile
}
\seealso{
Other sample-data: \code{\link{gt3x_dataurl}},
  \code{\link{gt3x_url}}
}
\concept{sample-data}
