% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_gt3x}
\alias{list_gt3x}
\title{List full paths to all gt3x files in a directory}
\usage{
list_gt3x(dirpath)
}
\description{
List full paths to all gt3x files in a directory
}
\examples{
list_gt3x(gt3x_datapath())

}
\seealso{
Other file manipulations: \code{\link{unzip.gt3x}}
}
\concept{file manipulations}
