% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_gt3x}
\alias{list_gt3x}
\title{List full paths to all gt3x files in a directory}
\usage{
list_gt3x(path)
}
\arguments{
\item{path}{Path(s) to file(s)}
}
\description{
List full paths to all gt3x files in a directory
}
\examples{
path <-
  system.file(
    "extdata",
    package = "read.gt3x")
list_gt3x(path)
\dontrun{
list_gt3x(gt3x_datapath())
}
}
\seealso{
Other file manipulations: 
\code{\link{gt3x_datapath}()},
\code{\link{is_gt3x}()},
\code{\link{unzip.gt3x}()}
}
\concept{file manipulations}
