% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGT3X.R
\name{parse_gt3x_info}
\alias{parse_gt3x_info}
\title{Parse GT3X info.txt file}
\usage{
parse_gt3x_info(path)
}
\arguments{
\item{path}{Path to a .gt3x file or an unzipped gt3x directory}
}
\description{
Parse GT3X info.txt file
}
\examples{
gt3xfile <- gt3x_datapath(1)
parse_gt3x_info(gt3xfile)

}
\concept{gt3x-parsers}
