% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_info.R
\name{print.gt3x_info}
\alias{print.gt3x_info}
\title{Print the contents of the info.txt file in a gt3x folder}
\usage{
\method{print}{gt3x_info}(x, ...)
}
\arguments{
\item{x}{gt3x_info object returned by parse_gt3x_info()}

\item{...}{not used}
}
\description{
Print the contents of the info.txt file in a gt3x folder
}
\seealso{
Other gt3x-parsers: 
\code{\link{as.data.frame.activity}()},
\code{\link{parse_gt3x_info}()},
\code{\link{read.gt3x}()}
}
\concept{gt3x-parsers}
