% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGT3X.R
\name{read.gt3x}
\alias{read.gt3x}
\title{Read GT3X}
\usage{
read.gt3x(path, verbose = FALSE, asDataFrame = FALSE, ...)
}
\arguments{
\item{path}{Path to gt3x folder}

\item{asDataFrame}{convert to an activity_df, see \code{as.data.frame.activity}}
}
\description{
Read activity samples from a GT3X file as a matrix
}
\examples{

# first unzip, then read
datadir <- gt3x_datapath()
gt3xfolders <- unzip.gt3x(datadir)
x <- read.gt3x(gt3xfolders[1])
df <- as.data.frame(x)
head(df)

# temporary unzip and read
gt3xfile <- gt3x_datapath(1)
x <- read.gt3x(gt3xfile)
df <- as.data.frame(x)
head(df)

}
\seealso{
Other gt3x-parsers: \code{\link{as.data.frame.activity}},
  \code{\link{parse_gt3x_info}},
  \code{\link{print.gt3x_info}}
}
\concept{gt3x-parsers}
