% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGT3X.R
\name{read.gt3x}
\alias{read.gt3x}
\title{Read GT3X}
\usage{
read.gt3x(path, verbose = FALSE, asDataFrame = FALSE,
  imputeZeroes = FALSE, ...)
}
\arguments{
\item{path}{Path to gt3x folder}

\item{verbose}{print diagnostic messages}

\item{asDataFrame}{convert to an activity_df, see \code{as.data.frame.activity}}

\item{imputeZeroes}{Impute zeros in case there are missingness? Default is FALSE, in which case
the time series will be incomplete in case there is missingness.}

\item{...}{additional arguments to pass to \code{parseGT3X} C++ code}
}
\value{
A numeric matrix with 3 columns (X, Y, Z) and the following attributes:
 \itemize{
\item \code{start_time} :  Start time from info file in \code{POSIXct} format.
\item \code{subject_name} : Subject name from info file
\item \code{time_zone} : Time zone from info file
\item \code{missingness} : Named integer vector. Names are Posixct timestamps and values are the number of missing values.
}
}
\description{
Read activity samples from a GT3X file as a matrix.
Please note that all timestams are in local time (of the device) even though they are
represented as POSIXct with GMT timezone.
}
\note{
The timestamps in the .gt3x data format are saved in .NET format, which is nanoseconds in local time since 0001-01-01.
This is a bit tricky to parse into an R datetime format. DateTimes are therefore represented as POSIXct format with the 'GMT' timezone attribute, which is false; the datetime actually
represents local time.
}
\examples{

# first unzip, then read
datadir <- gt3x_datapath()
gt3xfolders <- unzip.gt3x(datadir)
gt3xfile <- gt3xfolders[2]
x <- read.gt3x(gt3xfile, imputeZeroes = T)
df2 <- as.data.frame(x)
head(df2)

# temporary unzip, read, convert to a data frame
gt3xfile <- gt3x_datapath(1)
df <- read.gt3x(gt3xfile, asDataFrame = TRUE)
head(df)

}
\seealso{
Other gt3x-parsers: \code{\link{as.data.frame.activity}},
  \code{\link{parse_gt3x_info}},
  \code{\link{print.gt3x_info}}
}
\concept{gt3x-parsers}
