% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ticks2datetime}
\alias{ticks2datetime}
\title{Convert NET ticks to POSIXct datetime}
\usage{
ticks2datetime(ticks, tz)
}
\description{
Convert NET ticks to POSIXct datetime
}
\details{
reference: \url{https://stackoverflow.com/questions/35240874/r-net-ticks-to-timestamp-in-r}
}
\seealso{
Other gt3x-utils: \code{\link{get_n_samples}},
  \code{\link{have_log_and_info}}
}
\concept{gt3x-utils}
