% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_unzip.R
\name{unzip.gt3x}
\alias{unzip.gt3x}
\title{Unzip gt3x files}
\usage{
unzip.gt3x(path, location = tempdir(), remove_original = FALSE)
}
\arguments{
\item{path}{One of the following: (1) A path to a directory with .gt3x files in which case they are all unzipped, or
(2) A character vector of direct paths to .gt3x files.}

\item{location}{Path to a directory to unzip the files to. Default is a temporary directory created by tempdir().}

\item{remove_original}{Remove the zipfiles after unzipping?}
}
\value{
Returns a vector of paths to unzipped gt3x folders.
}
\description{
unzip.gt3x() makes it convenient to unzip multiple .gt3x files.
}
\details{
A .gt3x file is a zipped directory with two files: log.bin and info.txt.
 This function simply unzips the contents of the directories.
}
\examples{

# unzip a single .gt3x file
path <- gt3x_datapath(1)
gt3xdir <- unzip.gt3x(path)

# unzip multiple .gt3x files
dir <- gt3x_datapath()
gt3xdirs <- unzip.gt3x(dir)

}
\seealso{
Other file manipulations: \code{\link{gt3x_datapath}},
  \code{\link{is_gt3x}}, \code{\link{list_gt3x}}
}
\concept{file manipulations}
